Adding error handling

+++ src/main/java/com/polizel/springbootreact/student/StudentService.java
@@ -1,5 +1,7 @@
 package com.polizel.springbootreact.student;
 
+import com.polizel.springbootreact.student.exception.BadRequestException;
+import com.polizel.springbootreact.student.exception.StudentNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -19,10 +21,17 @@ public class StudentService {
     }
 
     public void addStudent(Student student){
+        if(studentRepository.selectExistsEmail(student.getEmail())){
+            throw new BadRequestException("Email " + student.getEmail() + " taken");
+        }
         studentRepository.save(student);
     }
 
     public void removeStudent(Long studentId) {
+        if(!studentRepository.existsById(studentId)) {
+            throw new StudentNotFoundException(
+                    "Student with id " + studentId + " does not exists");
+        }
         studentRepository.deleteById(studentId);
     }
 }
